refactor: input, printer 패키지 view로 변경
