+++ src/main/java/racingcar/view/input/request/CarRequest.java
@@ -1,4 +1,4 @@
-package racingcar.input.request;
+package racingcar.view.input.request;
 
 import racingcar.domain.car.CarResource;
